About Me Section
Created two versions of the About Me section, commented out the code for Version A but kept as reference. Added a placeholder image for my profile picture. Attempted to use Californian FB font but could not due to not having a site to host the font itself, went with Red Hat display font on Google Fonts instead. Section is responsive.