Header Fade Effect
I felt that the header looked too static with just an image so I was looking to add in a fade-in text effect but scrapped it since I couldn't figure how to make it responsive. Same with a typewriter effect. So I settled on coding in a fade-in image animation where it fades from black to the image when you first visit the page. I quite like how clean it came out. I want to animate my profile picture now too.